bio-age-coach/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
streamlit>=1.30.0
matplotlib>=3.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
deepeval>=0.8.0
plotly>=5.18.0
pandas>=2.2.0 